Add HTTP Basic
Some of the tests relied on HTTP Basic authentication. However, this was
not enabled. The tests were verifying the status as 200, but that was
actually the result of the login page which TestRestTemplate saw since it
automatically followed the redirects.

We also update the tests to use the HTTP Status code now.

NOTE:

Spring Security does content negotiation when both Form and HTTP
Basic are enabled. It will perform HTTP Basic if it looks like a REST call
(i.e. application/json). If it does not look like a REST call, then form
log in will be used. The order matters for formLogin() and httpBasic() as
it impacts which will be used (401 or redirect to log in page) in the event
Spring Security cannot guess what to do.