Remove LoginService
With HTTP Basic enabled, we don't need the LoginService. Any REST client
(including JavaScript) can use HTTP Basic authentication.

NOTE: By default Spring Security will return a 401 with out the
WWW-Authenticate header to JavaScript clients. This is necessary since
the browser will not allow JavaScript to handle a response with the
WWW-Authenticate header. This is determined by detecting the precense of
X-Requested-With header.

Also of interest is that Spring Security can use the HttpServletRequest
as a login service too. This means that the JavaEELoginService could also
be used with Spring Security.